move table template selectors to proper dir
