+++ projects/ngb-table/src/lib/children-selectors/ngb-table-body-row.directive.ts
