+++ projects/ngb-table/src/lib/children-selectors/ngb-table-footer-row.directive.spec.ts
