+++ projects/ngb-table/src/lib/children-selectors/ngb-table-footer.directive.ts
