+++ projects/ngb-table/src/lib/children-selectors/ngb-table-head.directive.spec.ts
