+++ projects/ngb-table/src/lib/children-selectors/ngb-table-header.directive.spec.ts
