+++ projects/ngb-table/src/lib/ngb-table.module.ts
@@ -1,12 +1,11 @@
 import { NgModule } from '@angular/core';
-import { NgbTableComponent } from './ngb-table.component';
-import { NgbTableHeaderDirective } from './ngb-table-header.directive';
 import { BrowserModule } from '@angular/platform-browser';
-import { NgbTableFooterDirective } from './ngb-table-footer.directive';
-import { NgbTableFooterRowDirective } from './ngb-table-footer-row.directive';
-import { NgbTableHeadDirective } from './ngb-table-head.directive';
-import { NgbTableBodyRowDirective } from './ngb-table-body-row.directive';
-
+import { NgbTableComponent } from './ngb-table.component';
+import { NgbTableHeaderDirective } from './children-selectors/ngb-table-header.directive';
+import { NgbTableFooterDirective } from './children-selectors/ngb-table-footer.directive';
+import { NgbTableFooterRowDirective } from './children-selectors/ngb-table-footer-row.directive';
+import { NgbTableHeadDirective } from './children-selectors/ngb-table-head.directive';
+import { NgbTableBodyRowDirective } from './children-selectors/ngb-table-body-row.directive';
 
 
 @NgModule({
